Overhaul WIZnet W5500 passkey pattern keys
Resolves #2397 (Overhaul WIZnet W5500 passkey pattern keys).
