+++ source/CMakeLists.txt
@@ -67,8 +67,8 @@ set(
     "picolibrary/wiznet/w5500.cc"
     "picolibrary/wiznet/w5500/ip.cc"
     "picolibrary/wiznet/w5500/ip/network_stack.cc"
-    "picolibrary/wiznet/w5500/ip/network_stack_keys.cc"
     "picolibrary/wiznet/w5500/ip/tcp.cc"
+    "picolibrary/wiznet/w5500/keys.cc"
 )
 set( PICOLIBRARY_LINK_LIBRARIES )
